+++ src/assign/propagate.rs
@@ -17,7 +17,10 @@ pub trait PropagateIF {
     ///
     /// ## Warning
     /// Callers must assure the consistency after this assignment.
-    fn assign_by_implication(&mut self, l: Lit, lv: DecisionLevel, cid: ClauseId, by: Option<Lit>);
+    #[cfg(feature = "chrono_BT")]
+    fn assign_by_implication(&mut self, l: Lit, lv: DecisionLevel, cid: ClauseId, by: Lit);
+    #[cfg(not(feature = "chrono_BT"))]
+    fn assign_by_implication(&mut self, l: Lit, cid: ClauseId, by: Lit);
     /// unsafe assume (assign by decision); doesn't emit an exception.
     /// ## Caveat
     /// Callers have to assure the consistency after this assignment.
@@ -27,17 +30,11 @@ pub trait PropagateIF {
     /// execute backjump in vivification sandbox
     fn backtrack_sandbox(&mut self);
     /// execute *boolean constraint propagation* or *unit propagation*.
-    fn propagate<C>(&mut self, cdb: &mut C) -> Option<ConflictContext>
-    where
-        C: ClauseDBIF;
+    fn propagate(&mut self, cdb: &mut impl ClauseDBIF) -> PropagationResult;
     /// `propagate` for vivification, which allows dead clauses.
-    fn propagate_sandbox<C>(&mut self, cdb: &mut C) -> Option<ClauseId>
-    where
-        C: ClauseDBIF;
+    fn propagate_sandbox(&mut self, cdb: &mut impl ClauseDBIF) -> PropagationResult;
     /// propagate then clear asserted literals
-    fn clear_asserted_literals<C>(&mut self, cdb: &mut C) -> MaybeInconsistent
-    where
-        C: ClauseDBIF;
+    fn clear_asserted_literals(&mut self, cdb: &mut impl ClauseDBIF) -> MaybeInconsistent;
 }
 
 #[cfg(feature = "unsafe_access")]
@@ -135,7 +132,8 @@ impl PropagateIF for AssignStack {
             _ => Err(SolverError::RootLevelConflict(Some(ClauseId::from(l)))),
         }
     }
-    fn assign_by_implication(&mut self, l: Lit, lv: DecisionLevel, cid: ClauseId, by: Option<Lit>) {
+    #[cfg(feature = "chrono_BT")]
+    fn assign_by_implication(&mut self, l: Lit, lv: DecisionLevel, cid: ClauseId, by: Lit) {
         debug_assert!(usize::from(l) != 0, "Null literal is about to be enqueued");
         debug_assert!(l.vi() < self.var.len());
         // The following doesn't hold anymore by using chronoBT.
@@ -150,7 +148,39 @@ impl PropagateIF for AssignStack {
         debug_assert!(self.trail.iter().all(|rl| *rl != l));
         set_assign!(self, l);
         self.level[vi] = lv;
-        self.reason[vi] = AssignReason::Implication(cid, by.unwrap_or(NULL_LIT));
+        self.reason[vi] = AssignReason::Implication(cid, by);
+        self.reward_at_assign(vi);
+        debug_assert!(!self.trail.contains(&l));
+        debug_assert!(!self.trail.contains(&!l));
+        self.trail.push(l);
+        if self.root_level == lv {
+            self.make_var_asserted(vi);
+        }
+
+        #[cfg(feature = "boundary_check")]
+        {
+            self.var[vi].propagated_at = self.num_conflict;
+            self.var[vi].state = VarState::Assigned(self.num_conflict);
+        }
+    }
+    #[cfg(not(feature = "chrono_BT"))]
+    fn assign_by_implication(&mut self, l: Lit, cid: ClauseId, by: Lit) {
+        debug_assert!(usize::from(l) != 0, "Null literal is about to be enqueued");
+        debug_assert!(l.vi() < self.var.len());
+        // The following doesn't hold anymore by using chronoBT.
+        // assert!(self.trail_lim.is_empty() || !cid.is_none());
+        let vi = l.vi();
+        debug_assert!(!self.var[vi].is(Flag::ELIMINATED));
+        debug_assert!(
+            var_assign!(self, vi) == Some(bool::from(l)) || var_assign!(self, vi).is_none()
+        );
+        debug_assert_eq!(self.assign[vi], None);
+        debug_assert_eq!(self.reason[vi], AssignReason::None);
+        debug_assert!(self.trail.iter().all(|rl| *rl != l));
+        set_assign!(self, l);
+        let lv = self.decision_level();
+        self.level[vi] = lv;
+        self.reason[vi] = AssignReason::Implication(cid, by);
         self.reward_at_assign(vi);
         debug_assert!(!self.trail.contains(&l));
         debug_assert!(!self.trail.contains(&!l));
@@ -198,11 +228,15 @@ impl PropagateIF for AssignStack {
             self.save_best_phases();
             self.best_assign = false;
         }
-        // We assume that backtrack is never happened in level zero.
-        let lim = self.trail_lim[lv as usize];
 
         #[cfg(feature = "chrono_BT")]
         let mut unpropagated: Vec<Lit> = Vec::new();
+
+        // We assume that backtrack is never happened in level zero.
+        let lim = self.trail_lim[lv as usize];
+
+        #[cfg(feature = "trail_saving")]
+        self.save_trail(lv);
 
         for i in lim..self.trail.len() {
             let l = self.trail[i];
@@ -243,8 +277,12 @@ impl PropagateIF for AssignStack {
 
             unset_assign!(self, vi);
             self.reason[vi] = AssignReason::None;
-            self.reward_at_unassign(vi);
-            self.insert_heap(vi);
+
+            #[cfg(not(feature = "trail_saving"))]
+            {
+                self.reward_at_unassign(vi);
+                self.insert_heap(vi);
+            }
         }
         self.trail.truncate(lim);
         // moved below -- self.q_head = self.trail.len();
@@ -296,11 +334,19 @@ impl PropagateIF for AssignStack {
     ///    So Eliminator should call `garbage_collect` before me.
     ///  - The order of literals in binary clauses will be modified to hold
     ///    propagation order.
-    fn propagate<C>(&mut self, cdb: &mut C) -> Option<ConflictContext>
-    where
-        C: ClauseDBIF,
-    {
+    fn propagate(&mut self, cdb: &mut impl ClauseDBIF) -> PropagationResult {
         let dl = self.decision_level();
+
+        #[cfg(feature = "trail_saving")]
+        if let cc @ Err(_) = self.from_saved_trail(cdb) {
+            self.num_propagation += 1;
+            self.num_conflict += 1;
+            self.num_reconflict += 1;
+            self.dpc_ema.update(self.num_decision);
+            self.ppc_ema.update(self.num_propagation);
+            return cc;
+        }
+
         while let Some(p) = self.trail.get(self.q_head) {
             self.num_propagation += 1;
             self.q_head += 1;
@@ -345,16 +391,19 @@ impl PropagateIF for AssignStack {
                             cdb[cid].moved_at = Propagate::EmitConflict(self.num_conflict, blocker);
                         }
 
-                        return Some(ConflictContext { cid, link: blocker });
+                        return Err(ConflictContext { cid, link: blocker });
                     }
                     None => {
                         debug_assert!(cdb[cid].lit0() == false_lit || cdb[cid].lit1() == false_lit);
+                        #[cfg(feature = "chrono_BT")]
                         self.assign_by_implication(
                             blocker,
                             self.level[propagating.vi()],
                             cid,
-                            Some(propagating),
+                            propagating,
                         );
+                        #[cfg(not(feature = "chrono_BT"))]
+                        self.assign_by_implication(blocker, cid, propagating);
                     }
                 }
             }
@@ -500,11 +549,12 @@ impl PropagateIF for AssignStack {
                         cdb[cid].moved_at = Propagate::EmitConflict(self.num_conflict, cached);
                     }
 
-                    return Some(ConflictContext {
+                    return Err(ConflictContext {
                         cid,
                         link: NULL_LIT,
                     });
                 }
+
                 #[cfg(feature = "chrono_BT")]
                 let dl = cdb[cid]
                     .iter()
@@ -512,15 +562,29 @@ impl PropagateIF for AssignStack {
                     .map(|l| self.level[l.vi()])
                     .max()
                     .unwrap_or(self.root_level);
+
                 debug_assert_eq!(cdb[cid].lit0(), cached);
                 debug_assert_eq!(self.assigned(cached), None);
-                assert!(other_watch_value.is_none());
-                self.assign_by_implication(cached, dl, cid, None);
+                debug_assert!(other_watch_value.is_none());
+
+                #[cfg(feature = "chrono_BT")]
+                self.assign_by_implication(cached, dl, cid, NULL_LIT);
+                #[cfg(not(feature = "chrono_BT"))]
+                self.assign_by_implication(cached, cid, NULL_LIT);
 
                 #[cfg(feature = "boundary_check")]
                 {
                     cdb[cid].moved_at = Propagate::BecameUnit(self.num_conflict, cached);
                 }
+            }
+            #[cfg(feature = "trail_saving")]
+            if let cc @ Err(_) = self.from_saved_trail(cdb) {
+                self.num_propagation += 1;
+                self.num_conflict += 1;
+                self.num_reconflict += 1;
+                self.dpc_ema.update(self.num_decision);
+                self.ppc_ema.update(self.num_propagation);
+                return cc;
             }
         }
         let na = self.q_head + self.num_eliminated_vars + self.num_asserted_vars;
@@ -528,7 +592,7 @@ impl PropagateIF for AssignStack {
             self.best_assign = true;
             self.num_best_assign = na;
         }
-        None
+        Ok(())
     }
     //
     //## How to generate propagate_sandbox from propagate
@@ -542,11 +606,7 @@ impl PropagateIF for AssignStack {
     // 1. (allow dead clauses)
     // 1. (allow eliminated vars)
     //
-    fn propagate_sandbox<C>(&mut self, cdb: &mut C) -> Option<ClauseId>
-    where
-        C: ClauseDBIF,
-    {
-        let dl = self.decision_level();
+    fn propagate_sandbox(&mut self, cdb: &mut impl ClauseDBIF) -> PropagationResult {
         while let Some(p) = self.trail.get(self.q_head) {
             self.q_head += 1;
             #[cfg(feature = "debug_propagation")]
@@ -576,15 +636,19 @@ impl PropagateIF for AssignStack {
 
                 match lit_assign!(self, blocker) {
                     Some(true) => (),
-                    Some(false) => return Some(cid),
+                    Some(false) => return Err(ConflictContext { cid, link: blocker }),
                     None => {
                         debug_assert!(cdb[cid].lit0() == false_lit || cdb[cid].lit1() == false_lit);
+
+                        #[cfg(feature = "chrono_BT")]
                         self.assign_by_implication(
                             blocker,
                             self.level[false_lit.vi()],
                             cid,
-                            Some(propagating),
+                            propagating,
                         );
+                        #[cfg(not(feature = "chrono_BT"))]
+                        self.assign_by_implication(blocker, cid, propagating);
                     }
                 }
             }
@@ -702,7 +766,10 @@ impl PropagateIF for AssignStack {
                             Propagate::SandboxEmitConflict(self.num_conflict, propagating);
                     }
 
-                    return Some(cid);
+                    return Err(ConflictContext {
+                        cid,
+                        link: NULL_LIT,
+                    });
                 }
 
                 #[cfg(feature = "chrono_BT")]
@@ -714,8 +781,12 @@ impl PropagateIF for AssignStack {
                     .unwrap_or(self.root_level);
                 debug_assert_eq!(cdb[cid].lit0(), cached);
                 debug_assert_eq!(self.assigned(cached), None);
-                assert!(other_watch_value.is_none());
-                self.assign_by_implication(cached, dl, cid, None);
+                debug_assert!(other_watch_value.is_none());
+
+                #[cfg(feature = "chrono_BT")]
+                self.assign_by_implication(cached, dl, cid, NULL_LIT);
+                #[cfg(not(feature = "chrono_BT"))]
+                self.assign_by_implication(cached, cid, NULL_LIT);
 
                 #[cfg(feature = "boundary_check")]
                 {
@@ -723,25 +794,20 @@ impl PropagateIF for AssignStack {
                 }
             }
         }
-        None
+        Ok(())
     }
-    fn clear_asserted_literals<C>(&mut self, cdb: &mut C) -> MaybeInconsistent
-    where
-        C: ClauseDBIF,
-    {
+    fn clear_asserted_literals(&mut self, cdb: &mut impl ClauseDBIF) -> MaybeInconsistent {
         assert_eq!(self.decision_level(), self.root_level);
         loop {
             if self.remains() {
-                self.propagate(cdb).map_or(Ok(()), |cc| {
-                    Err(SolverError::RootLevelConflict(Some(cc.cid)))
-                })?;
+                self.propagate_sandbox(cdb)
+                    .map_err(|cc| SolverError::RootLevelConflict(Some(cc.cid)))?;
             }
             self.propagate_at_root_level(cdb)
                 .map_or(Ok(()), |cc| Err(SolverError::RootLevelConflict(Some(cc))))?;
             if self.remains() {
-                self.propagate(cdb).map_or(Ok(()), |cc| {
-                    Err(SolverError::RootLevelConflict(Some(cc.cid)))
-                })?;
+                self.propagate_sandbox(cdb)
+                    .map_err(|cc| SolverError::RootLevelConflict(Some(cc.cid)))?;
             } else {
                 break;
             }
@@ -760,10 +826,7 @@ impl AssignStack {
         assert_ne!(self.assigned(b1), Some(false));
     }
     ///
-    fn propagate_at_root_level<C>(&mut self, cdb: &mut C) -> Option<ClauseId>
-    where
-        C: ClauseDBIF,
-    {
+    fn propagate_at_root_level(&mut self, cdb: &mut impl ClauseDBIF) -> Option<ClauseId> {
         let mut num_propagated = 0;
         while num_propagated < self.trail.len() {
             num_propagated = self.trail.len();
@@ -781,12 +844,12 @@ impl AssignStack {
                     RefClause::EmptyClause => return Some(cid),
                     RefClause::RegisteredClause(_) => (),
                     RefClause::UnitClause(lit) => {
-                        assert!(self.assigned(lit).is_none());
+                        debug_assert!(self.assigned(lit).is_none());
                         cdb.certificate_add_assertion(lit);
                         if self.assign_at_root_level(lit).is_err() {
                             return Some(cid);
                         } else {
-                            assert!(!self.locked(&cdb[cid], cid));
+                            debug_assert!(!self.locked(&cdb[cid], cid));
                             cdb.remove_clause(cid);
                         }
                     }