+++ src/processor/heap.rs
@@ -5,18 +5,12 @@ use {
 };
 
 pub trait VarOrderIF {
-    fn clear<A>(&mut self, asg: &mut A)
-    where
-        A: AssignIF;
+    fn clear(&mut self, asg: &mut impl AssignIF);
     fn len(&self) -> usize;
     fn insert(&mut self, occur: &[LitOccurs], vi: VarId, upward: bool);
     fn is_empty(&self) -> bool;
-    fn select_var<A>(&mut self, occur: &[LitOccurs], asg: &A) -> Option<VarId>
-    where
-        A: AssignIF;
-    fn rebuild<A>(&mut self, asg: &A, occur: &[LitOccurs])
-    where
-        A: AssignIF;
+    fn select_var(&mut self, occur: &[LitOccurs], asg: &impl AssignIF) -> Option<VarId>;
+    fn rebuild(&mut self, asg: &impl AssignIF, occur: &[LitOccurs]);
 }
 
 /// Mapping from Literal to Clauses.
@@ -113,10 +107,7 @@ impl VarOrderIF for VarOccHeap {
         self.idxs[0] = n;
         self.percolate_up(occur, n);
     }
-    fn clear<A>(&mut self, asg: &mut A)
-    where
-        A: AssignIF,
-    {
+    fn clear(&mut self, asg: &mut impl AssignIF) {
         for v in &mut self.heap[0..self.idxs[0] as usize] {
             asg.var_mut(*v as usize).turn_off(Flag::ENQUEUED);
         }
@@ -128,10 +119,7 @@ impl VarOrderIF for VarOccHeap {
     fn is_empty(&self) -> bool {
         self.idxs[0] == 0
     }
-    fn select_var<A>(&mut self, occur: &[LitOccurs], asg: &A) -> Option<VarId>
-    where
-        A: AssignIF,
-    {
+    fn select_var(&mut self, occur: &[LitOccurs], asg: &impl AssignIF) -> Option<VarId> {
         loop {
             let vi = self.get_root(occur);
             if vi == 0 {
@@ -142,10 +130,7 @@ impl VarOrderIF for VarOccHeap {
             }
         }
     }
-    fn rebuild<A>(&mut self, asg: &A, occur: &[LitOccurs])
-    where
-        A: AssignIF,
-    {
+    fn rebuild(&mut self, asg: &impl AssignIF, occur: &[LitOccurs]) {
         self.reset();
         for (vi, v) in asg.var_iter().enumerate().skip(1) {
             if asg.assign(vi).is_none() && !v.is(Flag::ELIMINATED) {