+++ src/solver/conflict.rs
@@ -170,7 +170,12 @@ pub fn handle_conflict(
             debug_assert_eq!(l1, cdb[cid].lit1());
             debug_assert_eq!(asg.assigned(l1), Some(false));
             debug_assert_eq!(asg.assigned(l0), None);
-            asg.assign_by_implication(l0, assign_level, cid, Some(!l1));
+
+            #[cfg(feature = "chrono_BT")]
+            asg.assign_by_implication(l0, assign_level, cid, !l1);
+            #[cfg(not(feature = "chrono_BT"))]
+            asg.assign_by_implication(l0, cid, !l1);
+
             // || check_graph(asg, cdb, l0, "biclause");
             rst.update(ProgressUpdate::LBD(1));
             for cid in &state.derive20 {
@@ -185,7 +190,12 @@ pub fn handle_conflict(
 
             debug_assert_eq!(cdb[cid].lit0(), l0);
             debug_assert_eq!(asg.assigned(l0), None);
-            asg.assign_by_implication(l0, assign_level, cid, None);
+
+            #[cfg(feature = "chrono_BT")]
+            asg.assign_by_implication(l0, assign_level, cid, NULL_LIT);
+            #[cfg(not(feature = "chrono_BT"))]
+            asg.assign_by_implication(l0, cid, NULL_LIT);
+
             // || check_graph(asg, cdb, l0, "clause");
             let lbd = cdb[cid].rank;
             rst.update(ProgressUpdate::LBD(lbd));
@@ -205,7 +215,12 @@ pub fn handle_conflict(
             );
             debug_assert_eq!(asg.assigned(l1), Some(false));
             debug_assert_eq!(asg.assigned(l0), None);
-            asg.assign_by_implication(l0, assign_level, cid, Some(!l1));
+
+            #[cfg(feature = "chrono_BT")]
+            asg.assign_by_implication(l0, assign_level, cid, !l1);
+            #[cfg(not(feature = "chrono_BT"))]
+            asg.assign_by_implication(l0, cid, !l1);
+
             // || check_graph(asg, cdb, l0, "registeredclause");
         }
         RefClause::UnitClause(_) => panic!("impossible"),
@@ -374,8 +389,8 @@ fn conflict_analyze(
                         debug_assert!(!asg.var(vi).is(Flag::ELIMINATED));
                         debug_assert!(
                             asg.assign(vi).is_some(),
-                            "conflict_analysis found {} {}",
-                            asg.var(vi),
+                            "conflict_analysis found V{} {}",
+                            vi,
                             asg.reason(vi),
                         );
                         debug_assert!(lvl <= dl);